+++ tests/test.py
@@ -1,10 +1,11 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-# Copyright (c) 2013, 2016 Yu-Jie Lin
+# Copyright (c) 2017 Leo Hemsted
 # Licensed under the BSD License, for detailed license information, see COPYING
 
 import doctest
 import unittest
+from time import time
 
 import smartypants
 from smartypants import smartypants as sp
@@ -146,6 +147,16 @@ document.write('<a href="' + href + '">' + linktext + "</a>");
 
         self.assertEqual(sp('"quote here"', Attr.set1 | Attr.s),
                          '"quote here"')
+
+    def test_incredibly_long_string(self):
+        # make sure it doesn't take too long with a long string
+        start = time()
+
+        sp("a" * 10000)
+
+        end = time()
+
+        self.assertLess(end - start, 0.2)
 
 
 def load_tests(loader, tests, pattern):